third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/rendering/widgets/baseline-alignment-and-overflow.tentative-expected.txt

This is a testharness.js-based test.
Found 132 tests; 84 PASS, 48 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS <input type="text" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="text" value="x" style="overflow: hidden; appearance: auto;">
PASS <input type="text" value="x" style="overflow: scroll; appearance: auto;">
PASS <input type="text" value="x" style="overflow: visible; appearance: none;">
PASS <input type="text" value="x" style="overflow: hidden; appearance: none;">
PASS <input type="text" value="x" style="overflow: scroll; appearance: none;">
PASS <input type="search" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="search" value="x" style="overflow: hidden; appearance: auto;">
PASS <input type="search" value="x" style="overflow: scroll; appearance: auto;">
PASS <input type="search" value="x" style="overflow: visible; appearance: none;">
PASS <input type="search" value="x" style="overflow: hidden; appearance: none;">
PASS <input type="search" value="x" style="overflow: scroll; appearance: none;">
PASS <input type="tel" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="tel" value="x" style="overflow: hidden; appearance: auto;">
PASS <input type="tel" value="x" style="overflow: scroll; appearance: auto;">
PASS <input type="tel" value="x" style="overflow: visible; appearance: none;">
PASS <input type="tel" value="x" style="overflow: hidden; appearance: none;">
PASS <input type="tel" value="x" style="overflow: scroll; appearance: none;">
PASS <input type="url" value="data:,x" style="overflow: visible; appearance: auto;">
PASS <input type="url" value="data:,x" style="overflow: hidden; appearance: auto;">
PASS <input type="url" value="data:,x" style="overflow: scroll; appearance: auto;">
PASS <input type="url" value="data:,x" style="overflow: visible; appearance: none;">
PASS <input type="url" value="data:,x" style="overflow: hidden; appearance: none;">
PASS <input type="url" value="data:,x" style="overflow: scroll; appearance: none;">
PASS <input type="email" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="email" value="x" style="overflow: hidden; appearance: auto;">
PASS <input type="email" value="x" style="overflow: scroll; appearance: auto;">
PASS <input type="email" value="x" style="overflow: visible; appearance: none;">
PASS <input type="email" value="x" style="overflow: hidden; appearance: none;">
PASS <input type="email" value="x" style="overflow: scroll; appearance: none;">
PASS <input type="password" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="password" value="x" style="overflow: hidden; appearance: auto;">
PASS <input type="password" value="x" style="overflow: scroll; appearance: auto;">
PASS <input type="password" value="x" style="overflow: visible; appearance: none;">
PASS <input type="password" value="x" style="overflow: hidden; appearance: none;">
PASS <input type="password" value="x" style="overflow: scroll; appearance: none;">
FAIL <input type="date" value="2020-01-01" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="date" value="2020-01-01" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="date" value="2020-01-01" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="date" value="2020-01-01" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="date" value="2020-01-01" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="date" value="2020-01-01" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="month" value="2020-01" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="week" value="2020-W01" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="time" value="00:00" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="time" value="00:00" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="time" value="00:00" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="time" value="00:00" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="time" value="00:00" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="time" value="00:00" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 24
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
FAIL <input type="datetime-local" value="2020-01-01T00:00" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 23
PASS <input type="number" value="0" style="overflow: visible; appearance: auto;">
PASS <input type="number" value="0" style="overflow: hidden; appearance: auto;">
PASS <input type="number" value="0" style="overflow: scroll; appearance: auto;">
PASS <input type="number" value="0" style="overflow: visible; appearance: none;">
PASS <input type="number" value="0" style="overflow: hidden; appearance: none;">
PASS <input type="number" value="0" style="overflow: scroll; appearance: none;">
FAIL <input type="range" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="range" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="range" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="range" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="range" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="range" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 70 but got 60
FAIL <input type="color" value="#000000" style="overflow: visible; appearance: auto;"> assert_equals: <span>.offsetTop expected 48 but got 45
FAIL <input type="color" value="#000000" style="overflow: hidden; appearance: auto;"> assert_equals: <span>.offsetTop expected 48 but got 70
FAIL <input type="color" value="#000000" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 48 but got 70
FAIL <input type="color" value="#000000" style="overflow: visible; appearance: none;"> assert_equals: <span>.offsetTop expected 48 but got 45
FAIL <input type="color" value="#000000" style="overflow: hidden; appearance: none;"> assert_equals: <span>.offsetTop expected 48 but got 70
FAIL <input type="color" value="#000000" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 48 but got 70
PASS <input type="checkbox" style="overflow: visible; appearance: auto;">
PASS <input type="checkbox" style="overflow: hidden; appearance: auto;">
PASS <input type="checkbox" style="overflow: scroll; appearance: auto;">
PASS <input type="checkbox" style="overflow: visible; appearance: none;">
PASS <input type="checkbox" style="overflow: hidden; appearance: none;">
PASS <input type="checkbox" style="overflow: scroll; appearance: none;">
PASS <input type="radio" style="overflow: visible; appearance: auto;">
PASS <input type="radio" style="overflow: hidden; appearance: auto;">
PASS <input type="radio" style="overflow: scroll; appearance: auto;">
PASS <input type="radio" style="overflow: visible; appearance: none;">
PASS <input type="radio" style="overflow: hidden; appearance: none;">
PASS <input type="radio" style="overflow: scroll; appearance: none;">
PASS <input type="file" style="overflow: visible; appearance: auto;">
PASS <input type="file" style="overflow: hidden; appearance: auto;">
PASS <input type="file" style="overflow: scroll; appearance: auto;">
PASS <input type="file" style="overflow: visible; appearance: none;">
PASS <input type="file" style="overflow: hidden; appearance: none;">
PASS <input type="file" style="overflow: scroll; appearance: none;">
PASS <input type="submit" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="submit" value="x" style="overflow: hidden; appearance: auto;">
FAIL <input type="submit" value="x" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 27
PASS <input type="submit" value="x" style="overflow: visible; appearance: none;">
PASS <input type="submit" value="x" style="overflow: hidden; appearance: none;">
FAIL <input type="submit" value="x" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 27
PASS <input type="image" src="data:,broken" alt="x" style="overflow: visible; appearance: auto;">
PASS <input type="image" src="data:,broken" alt="x" style="overflow: hidden; appearance: auto;">
PASS <input type="image" src="data:,broken" alt="x" style="overflow: scroll; appearance: auto;">
PASS <input type="image" src="data:,broken" alt="x" style="overflow: visible; appearance: none;">
PASS <input type="image" src="data:,broken" alt="x" style="overflow: hidden; appearance: none;">
PASS <input type="image" src="data:,broken" alt="x" style="overflow: scroll; appearance: none;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: visible; appearance: auto;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: hidden; appearance: auto;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: scroll; appearance: auto;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: visible; appearance: none;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: hidden; appearance: none;">
PASS <input type="image" src="data:(png)" alt="x" style="overflow: scroll; appearance: none;">
PASS <input type="reset" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="reset" value="x" style="overflow: hidden; appearance: auto;">
FAIL <input type="reset" value="x" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 27
PASS <input type="reset" value="x" style="overflow: visible; appearance: none;">
PASS <input type="reset" value="x" style="overflow: hidden; appearance: none;">
FAIL <input type="reset" value="x" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 27
PASS <input type="button" value="x" style="overflow: visible; appearance: auto;">
PASS <input type="button" value="x" style="overflow: hidden; appearance: auto;">
FAIL <input type="button" value="x" style="overflow: scroll; appearance: auto;"> assert_equals: <span>.offsetTop expected 34 but got 27
PASS <input type="button" value="x" style="overflow: visible; appearance: none;">
PASS <input type="button" value="x" style="overflow: hidden; appearance: none;">
FAIL <input type="button" value="x" style="overflow: scroll; appearance: none;"> assert_equals: <span>.offsetTop expected 34 but got 27
Harness: the test ran to completion.
